+++ tools/vignette.R
@@ -1,6 +1,6 @@
 rm(list=ls())
 devtools::load_all()
-rmarkdown::render("vignettes/SM01-HIV.Rmd")
+rmarkdown::render("vignettes/DT00-DecisionTreeTutorial.Rmd")
 rm(list=ls())